separate classes to new files | handle new bucket name better
