+++ aws-s3-bucket-tool.rb
@@ -1,206 +1,11 @@
 #!/usr/bin/env ruby
-
+require_relative "ui-helper"
+require_relative "s3-helper"
 require 'aws-sdk-s3'
 require 'curses'
 
 PAGE_SIZE = 20 # Number of items per page
 
-class UI
-  def initialize(window)
-    @window = window
-  end
-
-  # Prompt the user with a message & return their input.
-  def prompt_user(prompt)
-    @window.setpos(Curses.lines - 3, 0)
-    @window.addstr(prompt)
-
-    Curses.curs_set(1)
-    Curses.echo
-    input = @window.getstr.strip
-    Curses.noecho
-    Curses.curs_set(0)
-
-    input
-  end
-
-  # Display an error message above the input line.
-  def display_error(message)
-    @window.setpos(Curses.lines - 5, 0)
-    @window.addstr(message.ljust(Curses.cols))
-    @window.addstr("Press any key to continue.")
-    @window.refresh
-    @window.getch
-  end
-
-  # Clears the line(s)
-  def clear_lines(lines)
-    for line in lines
-      @window.setpos(line, 0)
-      @window.addstr(" " * Curses.cols) # Clear error line
-    end
-  end
-
-  # Render UI with pagination
-  def display_page(bucket, prefix, items, page)
-    @window.clear
-    start_index = page * PAGE_SIZE
-    end_index = [start_index + PAGE_SIZE, items.size].min
-    page_count = (items.size / PAGE_SIZE.to_f).ceil
-
-    # Display bucket path, contents, and pagination info
-    @window.setpos(0, 0)
-    @window.addstr("Current Bucket: #{bucket} /#{prefix} (Page #{page + 1} of #{page_count})\nContents:\n")
-    items[start_index...end_index].each_with_index { |item, index| @window.addstr("[#{start_index + index}] #{item}\n") }
-
-    # Display page navigation prompts
-    @window.addstr("\n")
-    @window.addstr(if page == 0 && page_count > 1
-                      "(Press 'F' for next page.)"
-                    elsif page + 1 == page_count && page_count > 1
-                      "(Press 'P' for previous page.)"
-                    elsif page_count > 1
-                      "(Press 'P' for previous page. Press 'F' for next page.)"
-                    end.to_s)
-
-    # Options bar
-    @window.setpos(Curses.lines - 1, 0)
-    @window.attron(Curses::A_REVERSE) do
-      @window.addstr("[H]: Help | [0-9]: Select item | [N]: New Bucket URL | [B]: Back | [P]: Prev Page | [F]: Next Page | [Q]: Quit".ljust(Curses.cols))
-    end
-    @window.refresh
-  end
-
-  # Display help information
-  def display_help
-    help_text = "
-      [0-9] to select item
-      [Q] to quit
-      [N] to change bucket
-      [B] to go back
-      [P] to prev page
-      [F] to next page
-      Press any key to continue..."
-    @window.setpos(Curses.lines - (help_text.lines.count + 3), 0)
-    @window.addstr(help_text.ljust(Curses.cols))
-    @window.refresh
-    @window.getch
-  end
-end
-
-class S3Browser
-  def initialize(s3, ui)
-    @s3 = s3
-    @ui = ui
-  end
-
-  # List objects in the current bucket folder.
-  def list_objects(bucket, prefix = "")
-    begin
-      response = @s3.list_objects_v2(bucket: bucket, prefix: prefix, delimiter: '/')
-      folders = response.common_prefixes.map { |prefix_obj| prefix_obj.prefix.gsub(prefix, '').chomp('/') + '/' }
-      files = response.contents.map { |obj| obj.key.gsub(prefix, "") }
-      files.reject! { |f| f.include?("/") && f != prefix }
-    rescue Aws::S3::Errors::NoSuchBucket
-      @ui.display_error("Bucket does not exist. Check your spelling and try again.")
-      return nil
-    rescue Aws::S3::Errors::AccessDenied
-      @ui.display_error("Access denied to the bucket. Check your permissions and try again.")
-      return nil
-    rescue Aws::S3::Errors::InvalidBucketName
-      @ui.display_error("The bucket name format is invalid. Please enter a valid bucket name.")
-      return nil
-    rescue Seahorse::Client::NetworkingError
-      @ui.display_error("Network error: Unable to connect to AWS S3. Check your connection and try again.")
-      return nil
-    rescue Aws::S3::Errors::RequestTimeout
-      @ui.display_error("Request timed out. The network may be slow or unresponsive. Try again later.")
-      return nil
-    rescue Aws::S3::Errors::Throttling
-      @ui.display_error("Rate limit exceeded. Please wait a moment and try again.")
-      return nil
-    rescue StandardError => e
-      @ui.display_error("An error occurred: #{e.message}")
-      return nil
-    end
-
-    folders + files
-  end
-
-  # Download a file and confirm to the user
-  def download_file(bucket, file_key)
-    default_name = File.basename(file_key)
-    prompt = "Enter a new name for the file or press Enter to keep '#{default_name}': "
-    download_name = @ui.prompt_user(prompt)
-    download_name = default_name if download_name.empty?
-
-    begin
-      @s3.get_object(response_target: download_name, bucket: bucket, key: file_key)
-      @ui.display_error("Downloaded '#{download_name}'.")
-    rescue Aws::S3::Errors::NoSuchKey
-      @ui.display_error("The file does not exist in the bucket.")
-    rescue Aws::S3::Errors::AccessDenied
-      @ui.display_error("Access denied. You do not have permission to download this file.")
-    rescue Seahorse::Client::NetworkingError
-      @ui.display_error("Network error: Unable to download file. Check your connection and try again.")
-    rescue StandardError => e
-      @ui.display_error("An unexpected error occurred: #{e.message}")
-    # ensure
-    #   @window.refresh
-    #   @window.getch
-    end
-  end
-
-  # Navigate within the bucket and interact with files
-  def navigate_bucket(bucket)
-    prefix = ""
-    history = []
-    page = 0
-
-    loop do
-      items = list_objects(bucket, prefix)
-      return if items.nil? # Exit if there was an error listing objects
-      @ui.display_page(bucket, prefix, items, page)
-      input = @ui.prompt_user("Input: ").downcase
-
-      case input
-      when "q" then break
-      when "h" then @ui.display_help
-      when "n"
-        bucket = @ui.prompt_user("Enter new bucket name: ")
-        prefix = ""
-        history.clear
-        page = 0
-      when "b"
-        prefix = history.pop || ""
-        page = 0
-      when "f"
-        page += 1 if (page + 1) * PAGE_SIZE < items.size
-      when "p"
-        page -= 1 if page > 0
-      when /^[0-9]+$/
-        index = input.to_i
-        if index >= 0 && index < items.size
-          selected = items[index]
-          new_prefix = "#{prefix}#{selected}".chomp('/')
-          if selected.end_with?("/") # Folder
-            history.push(prefix)
-            prefix = "#{new_prefix}/"
-            page = 0
-          else # File
-            download_file(bucket, new_prefix)
-          end
-        end
-      else
-        @ui.display_error("Invalid option. Press 'H' for help.")
-        # @window.setpos(Curses.lines - 4, 0)
-        # @window.addstr("Invalid option. Press 'H' for help.".ljust(Curses.cols))
-        # @window.refresh
-        # @window.getch
-      end
-    end
-  end
-end
 
 # Main function to initialize Curses and start navigation
 def main(s3)
@@ -210,20 +15,24 @@ def main(s3)
   begin
     window = Curses.stdscr
     window.clear
-    ui = UI.new(window)
-    browser = S3Browser.new(s3, ui)
+    ui = UI.new(window, PAGE_SIZE)
+    browser = S3Browser.new(s3, ui, PAGE_SIZE)
 
-    # Loop to prompt for a valid bucket name until successful
-    bucket = nil
+    # Prompt for a valid bucket name until one is provided
     loop do
-      ui.clear_lines(Curses.lines - 5 .. Curses.lines - 3)
-      bucket = ui.prompt_user("Enter the S3 bucket name: ")
+      ui.clear_error_message
+      bucket = ui.prompt_user("Enter the S3 bucket name: ").strip
+      next if bucket.empty? # Retry if input is blank
 
-      # Break if valid bucket, otherwise display error
-      break if !browser.list_objects(bucket).nil?
+      # Check if the bucket is valid by attempting to list objects
+      items = browser.list_objects(bucket)
+      if items
+        # Only proceed if the bucket is valid
+        result = browser.navigate_bucket(bucket)
+        break unless result == :restart # Exit unless :restart is returned
+      end
     end
 
-    browser.navigate_bucket(bucket)
   rescue Aws::Sigv4::Errors::MissingCredentialsError
     ui.display_error("AWS credentials not found. Please configure your AWS credentials.")
   rescue Aws::Errors::ServiceError => e
